refactor: 기존 테스트와 새로운 테스트 병합
