+++ backend/src/test/java/com/festago/student/application/StudentServiceTest.java
@@ -2,6 +2,7 @@ package com.festago.student.application;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
@@ -14,7 +15,9 @@ import com.festago.member.repository.MemberRepository;
 import com.festago.school.domain.School;
 import com.festago.school.repository.SchoolRepository;
 import com.festago.student.domain.StudentCode;
+import com.festago.student.domain.VerificationCode;
 import com.festago.student.dto.StudentSendMailRequest;
+import com.festago.student.dto.StudentVerificateRequest;
 import com.festago.student.infrastructure.MockMailClient;
 import com.festago.student.infrastructure.RandomVerificationCodeProvider;
 import com.festago.student.repository.StudentCodeRepository;
@@ -161,6 +164,61 @@ class StudentServiceTest {
         void 인증메일_전송() {
             assertThatNoException()
                 .isThrownBy(() -> studentService.sendVerificationMail(1L, request));
+        }
+    }
+
+    @Nested
+    class 학생_인증 {
+
+        @Test
+        void 이미_학생인증정보가_존재하면_예외() {
+            // given
+            Long memberId = 1L;
+            StudentVerificateRequest request = new StudentVerificateRequest("123456");
+            given(studentRepository.existsByMemberId(memberId))
+                .willReturn(true);
+
+            // when & then
+            assertThatThrownBy(() -> studentService.verificate(memberId, request))
+                .isInstanceOf(BadRequestException.class)
+                .hasMessage("이미 학교 인증이 완료된 사용자입니다.");
+        }
+
+        @Test
+        void 인증_코드가_존재하지_않으면_예외() {
+            // given
+            Long memberId = 1L;
+            StudentVerificateRequest request = new StudentVerificateRequest("123456");
+            given(memberRepository.findById(anyLong()))
+                .willReturn(Optional.of(MemberFixture.member().build()));
+            given(studentCodeRepository.findByCodeAndMember(any(), any()))
+                .willReturn(Optional.empty());
+
+            // when & then
+            assertThatThrownBy(() -> studentService.verificate(memberId, request))
+                .isInstanceOf(BadRequestException.class)
+                .hasMessage("올바르지 않은 학생 인증 코드입니다.");
+        }
+
+        @Test
+        void 성공() {
+            // given
+            Long memberId = 1L;
+            StudentVerificateRequest request = new StudentVerificateRequest("123456");
+            Member member = MemberFixture.member().build();
+            given(memberRepository.findById(anyLong()))
+                .willReturn(Optional.of(member));
+            given(studentCodeRepository.findByCodeAndMember(any(), any()))
+                .willReturn(Optional.of(new StudentCode(
+                    new VerificationCode("123456"),
+                    new School("snu.ac.kr", "서울대학교"),
+                    member,
+                    "ohs"
+                )));
+
+            // when & then
+            assertThatNoException()
+                .isThrownBy(() -> studentService.verificate(memberId, request));
         }
     }
 }
